sini175* refactoring useEffects:  joining Lobby and starting game via websockets
